Obtain the hive metastore and hbase token as well as hdfs token in AMDelegationToeknRenewer.
